+++ Backend/Controllers/groupController.js
@@ -45,6 +45,8 @@ export const createGroup = asyncHandler(async (req, res, next) => {
   });
 });
 
+//in this you can check id user is already in group or not
+//currently it giving error if you try to add same user in group
 export const addUserToGroup = asyncHandler(async (req, res, next) => {
   const { userId } = req.body || {};
   const { groupId } = req.params;
@@ -135,6 +137,8 @@ export const getExpensesOfGroup = asyncHandler(async (req, res, next) => {
 export const deleteGroup = asyncHandler(async (req, res, next) => {
   const { groupId } = req.params;
 
+  const user = req.user;
+
   const transaction = await sequelize.transaction();
 
   try {
@@ -144,6 +148,24 @@ export const deleteGroup = asyncHandler(async (req, res, next) => {
       return next(new AppError("No group found with provided ID", 404));
     }
 
+    const members = await groupMembers.findAll({
+      where: { groupId: group.groupId },
+      transaction,
+    });
+
+    const membersIds = members.map((m) => m.userId);
+
+    if (!membersIds.includes(user.userId)) {
+      await transaction.rollback();
+      return next(
+        new AppError(
+          "You are not authorized to delete this group as you are not the member of this group",
+          403
+        )
+      );
+    }
+
+    // deleting data related to the group from group member table
     await groupMembers.destroy({ where: { groupId }, transaction });
 
     const expensesData = await expenses.findAll({
@@ -152,6 +174,7 @@ export const deleteGroup = asyncHandler(async (req, res, next) => {
     });
     const expensesIds = expensesData.map((e) => e.expenseId);
 
+    // deleting data related to the group from expense split table
     if (expensesIds.length > 0) {
       await expenseSplit.destroy({
         where: { expenseId: expensesIds },
@@ -159,8 +182,10 @@ export const deleteGroup = asyncHandler(async (req, res, next) => {
       });
     }
 
+    //deleting data related to the group from expense table
     await expenses.destroy({ where: { groupId }, transaction });
 
+    //deleting group itself
     await group.destroy({ transaction }); // directly deleteing instance
 
     await transaction.commit();
@@ -174,4 +199,24 @@ export const deleteGroup = asyncHandler(async (req, res, next) => {
     console.error("Error deleting group:", error.message);
     return next(error);
   }
+});
+
+export const settleUp = asyncHandler(async (req, res, next) => {
+  const { groupId } = req.params;
+
+  const group = await groups.findByPk(groupId);
+
+  if (!group) {
+    return next(new AppError("No group found with provided ID", 404));
+  }
+
+  const groupExpenses = await expenses.findAll({ where: { groupId } });
+
+  const eid = groupExpenses.map((e) => e.expenseId);
+
+  const groupSPlit = await expenseSplit.findAll({ where: { expenseId: eid } });
+
+  res.status(200).json({
+    groupSPlit,
+  });
 });
